Feature: Create a Surface from a RawWindowHandle
